+++ apm-agent-plugins/apm-jms-plugin/apm-jms-plugin-base/src/main/java/co/elastic/apm/agent/jms/JmsMessageListenerInstrumentation.java
@@ -18,6 +18,8 @@
  */
 package co.elastic.apm.agent.jms;
 
+import co.elastic.apm.agent.configuration.MessagingConfiguration;
+import co.elastic.apm.agent.impl.ElasticApmTracer;
 import co.elastic.apm.agent.impl.GlobalTracer;
 import co.elastic.apm.agent.impl.transaction.Transaction;
 import co.elastic.apm.agent.sdk.logging.Logger;
@@ -54,13 +56,14 @@ public class JmsMessageListenerInstrumentation extends BaseJmsInstrumentation {
     public static final Logger logger = LoggerFactory.getLogger(JmsMessageListenerInstrumentation.class);
 
     @Nullable
-    private JmsConfiguration configuration;
+    private MessagingConfiguration configuration;
+
+    public JmsMessageListenerInstrumentation(ElasticApmTracer tracer){
+        this.configuration = tracer.getConfig(MessagingConfiguration.class);
+    }
 
     @Override
     public ElementMatcher<? super NamedElement> getTypeMatcherPreFilter() {
-        if (configuration == null) {
-            configuration = GlobalTracer.getTracerImpl().getConfig(JmsConfiguration.class);
-        }
 
         ElementMatcher.Junction<NamedElement> nameMatcher = nameContains("$") // inner classes
             .or(nameContains("/")) // lambdas
