--- apm-agent-plugins/apm-jms-plugin/apm-jms-plugin-base/src/main/resources/META-INF/services/org.stagemonitor.configuration.ConfigurationOptionProvider
@@ -1,1 +1,0 @@
-co.elastic.apm.agent.jms.JmsConfiguration
